requirements.txt
Django==4.1.5
Jinja2==3.1.2
MarkupSafe==2.1.1
PySocks==1.7.1
PyYAML==6.0
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.11.1
blpapi==3.19.3
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
django-cors-headers==3.14.0
et-xmlfile==1.1.0
free-proxy==1.1.1
frozendict==2.3.5
html5lib==1.1
idna==3.4
jieba==0.42.1
langdetect==1.0.9
lxml==4.9.2
multitasking==0.0.11
my-fake-useragent==0.2.0
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.1
openpyxl==3.1.1
pandas==1.5.2
pip==23.0.1
pycparser==2.21
pygtrans==1.4.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
render==1.0.0
requests==2.28.2
requests-futures==1.0.0
setuptools==65.5.1
six==1.16.0
snownlp==0.12.3
soupsieve==2.3.2.post1
sqlparse==0.4.3
torch==1.13.1
tqdm==4.64.1
typing-extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
webencodings==0.5.1
wheel==0.38.4
xlrd==2.0.1	
xlwt==1.3.0
yahooquery==2.3.0